+++ assets/images/Stages/Levels/Level1/Stage4/Background/Spec/Level1Stage4BackgroundBSpec.c
@@ -21,8 +21,8 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level1Stage4BackgroundTiles[];
-extern BYTE Level1Stage4BackgroundBMap[];
+extern uint32 Level1Stage4BackgroundTiles[];
+extern uint16 Level1Stage4BackgroundBMap[];
 extern CharSetSpec Level1Stage4BackgroundCharset;
